hiver_starter

Module prelude

预导入模块 - 包含所有常用类型 预导入模块 / Prelude Module

包含 Hiver Starter 最常用的类型和宏，方便一行导入。 Contains the most commonly used Hiver Starter types and macros for one-line imports.

使用方式 / Usage

use hiver_starter::prelude::*;

Re-exports

pub use crate::config::ConfigurationLoader;

pub use crate::config::ConfigurationProperties;

pub use crate::config::Environment;

pub use crate::core::ApplicationContext;

pub use crate::core::AutoConfiguration;

pub use crate::core::BeanDefinition;

pub use crate::core::ComponentRegistry;

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Request

HTTP Request HTTP 请求

Response

HTTP Response / HTTP 响应

Router

HTTP Router HTTP路由器

StatusCode

HTTP Status Codes HTTP 状态码

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Type Aliases

Body

Body type alias using FullBody by default. 默认使用 FullBody 的 Body 类型别名。

Result

Result<T, Error>

Attribute Macros

autowired

组件注解宏 Component annotation macros

bean

组件注解宏 Component annotation macros

cache_evict

组件注解宏 Component annotation macros

cache_put

组件注解宏 Component annotation macros

cacheable

组件注解宏 Component annotation macros

component

组件注解宏 Component annotation macros

config

组件注解宏 Component annotation macros

configuration

组件注解宏 Component annotation macros

controller

组件注解宏 Component annotation macros

delete

组件注解宏 Component annotation macros

get

组件注解宏 Component annotation macros

head

组件注解宏 Component annotation macros

hiver_main

应用主宏 - 标记应用程序入口点 Similar to Spring Boot’s @SpringBootApplication

options

组件注解宏 Component annotation macros

patch

组件注解宏 Component annotation macros

post

组件注解宏 Component annotation macros

pre_authorize

组件注解宏 Component annotation macros

put

组件注解宏 Component annotation macros

repository

组件注解宏 Component annotation macros

scheduled

组件注解宏 Component annotation macros

secured

组件注解宏 Component annotation macros

service

组件注解宏 Component annotation macros

trace

组件注解宏 Component annotation macros

transactional

组件注解宏 Component annotation macros

validated

组件注解宏 Component annotation macros

value

组件注解宏 Component annotation macros

Derive Macros

Deserialize

Serialize